Inbox now works with install_requires and tests pass.
Summary:
This is required for using inbox as a package. When installed with pip, setup.py will install the packages in setup.py. I've also added MANIFEST files to include the necessary static files that should be shipped with the package
Adding mdown files to source dist

Test Plan: Ran tests locally

Reviewers: emfree

Differential Revision: https://review.inboxapp.com/D1058